86c18c3jb Implement hook_civicrm_enable
This hook will intercept any contribution creations that are associated with
the 'Gift' financial type. Note that this financial type is not native to
CiviCRM; it needs to be created.

For every such contribution, the hook will result in a negative sign on each
of the following entries:

* `amount` in `civicrm_financial_item`
* `unit_price` and `line_total` in `civicrm_line_item`
* `total_amount` and `net_amount` in `civicrm_financial_trxn`

It is assumed that fees and taxes for these contributions are zero.